Feat: 배차 확정 기능 구현

+++ src/main/java/com/team2/finalproject/domain/dispatchdetail/model/dto/response/DispatchDetailResponse.java
@@ -30,7 +30,7 @@ public class DispatchDetailResponse {
     private String smPhoneNumber;
 
     @Schema(example = "80", description = "용적률")
-    private int floorAreaRatio;
+    private double floorAreaRatio;
 
     @Schema(example = "WING_BODY", description = "차종")
     private VehicleType vehicleType;
